% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_and_scale.R
\name{center_and_scale}
\alias{center_and_scale}
\title{Performs Column Wise Centering and Scaling of a Matrix}
\usage{
center_and_scale(matrix.dgCMatrix, center = FALSE, scale = FALSE)
}
\arguments{
\item{matrix.dgCMatrix}{<matrix> Input matrix, possibly in sparse format}

\item{center}{<bool> default FALSE; Whether to center the input matrix}

\item{scale}{<bool> default FALSE; Whether to scale the input matrix}
}
\value{
The input matrix in dgCMatrix sparse format after centering,
scaling, both or just as it is depending on the flag center and scale.
}
\description{
This function taken as input a sparse matrix, generally dgCMatrix, and may
center and scale it in a column-wise fashion, or simply return the matrix
as it is.
}
\examples{
#center_and_scale(SingleCellExperiment::logcounts(sce), center = TRUE, scale = TRUE)
}
