% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_default_name.R
\name{change_default_name}
\alias{change_default_name}
\title{Changes Default Result Name only if it is already present}
\usage{
change_default_name(result_name, name_list)
}
\arguments{
\item{result_name}{<character> default result_name.}

\item{name_list}{<list> named list which should not
contain the default result_name as one of the names.}
}
\value{
Either the same result_name <character > or a new result_name which
accounts for the presence of multiple result_names similar to the one passed
to it. example result_name2 or result_name.
}
\description{
This function changes the default result name if it is already present
in a named list.
}
\examples{
#result_name <- change_default_name('pca', SingleCellExperiment::reducedDimNames(sce))
}
