% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dir.R
\name{check_dir}
\alias{check_dir}
\title{Checks if a Directory Exists, if it does not it creates it}
\usage{
check_dir(directory2check.dir)
}
\arguments{
\item{directory2check.dir}{<character> path to the directory you want to
check}
}
\value{
either nothing, if the directory already exists, or it creates
the directory and prompt the user with a message of where the directory has
been created
}
\description{
This function taken as input a directory and checks if it exists, if it does
not it creates it.
}
\examples{
#check_dir('PhD')
}
