% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_aa.R
\name{decomp_aa}
\alias{decomp_aa}
\title{Decomposition Approach : Archetypal Analysis}
\usage{
decomp_aa(
  sce,
  n_components,
  reduction_method = "pca",
  assay = "logcounts",
  n_dimensions = 50,
  reduced_representation = NULL,
  delta = 0,
  conv_crit = 1e-04,
  parallel_type = "m",
  volume_ratio = "t_ratio",
  sample_proportion = NULL,
  method = "pcha",
  normalise_var = TRUE,
  boostrap_number = 20,
  result_name = "aa",
  envname = "r-decomp",
  return_model = FALSE,
  seed = 42,
  verbose = FALSE
)
}
\arguments{
\item{sce}{<SingleCellExperiment object> with a your data}

\item{n_components}{<integer> number of archetypes desired}

\item{reduction_method}{<character> default *pca*;
one of : *'pca'* (sparse pca), *'nmf'* (dense nmf) or *'none'*
specifying the type of dimensionality reduction to be performed prior to AA.}

\item{assay}{<character> default 'logcounts'; assay used as input for the
reduction_method. If reduction_method = 'none, it will be the input for AA.
Ignored if reduced_representation is specified}

\item{n_dimensions}{<integer> default 50; number of dimensions used
to perform the reduction_method; ignored if reduced_representation
is specified.}

\item{reduced_representation}{<character> default NULL; It specifies an
already existing compressed representation to be used prior to AA.
If specified, reduction_method, n_dimensions and assay will be ignored.}

\item{delta}{<integer> default 0;  Inflates the original polytope (simplex)
fit such that it may contain more points of the data set. (from ParetoTI)}

\item{conv_crit}{<double> default 1e-04; convergence criteria; a lower value
(like 1e-6) will result in higher accuracy costing a longer computational
time. (from ParetoTI)}

\item{parallel_type}{<character> default *m*; one of *s*, *m*, *cmq*.
Do we want to parallelize? then use *m* for multicore.
*s* means single core processing using lapply.
*m* means multi-core parallel procession using parLapply.
*cmq* means multi-node parallel processing on a computing cluster
using clustermq package. (from ParetoTI)}

\item{volume_ratio}{<character> default *t_ratio*;
*t_ratio* : defines the volume of the convex hull of the data and the
t-ratio.
*variance_ratio* : defined the ratio of  variance of archetype positions to
variance of data; .
*none* :  do nothing; .
Caution! Geometric figure should be at least simplex:
qhull algorithm will fail to find convex hull of flat 2D shapes
in 3D, 3D shapes in 4D and so on. So, for calculating this dimensions
are selected based order of rows in data. Makes sense for principal
components but not for original data. Caution 2!
Computation time and memory use increase very quickly with dimensions.
Do not use for more than 7-8 dimensions. (from ParetoTI)}

\item{sample_proportion}{<double or NULL> default NULL; either NULL or the
proportion of data set that should be included in each sample.
If NULL the polytope fitting algorithm is run n times on
the same data which is useful for evaluating how often the
algorithm gets stuck in local optima. (from ParetoTI)}

\item{method}{<character> default *pcha*; method for archetypal analysis:
*pcha* : Principal Convex Hull,
*kmeans* : KMeans
*louvain* : Louvain (should be avoided)}

\item{normalise_var}{<bool> default TRUE; either TRUE or FALSE,
whether to normalise variance in position of archetypes by variance
in data in each dimensions.(from ParetoTI).}

\item{boostrap_number}{<integer or NA> default 20; number of bootstrap
samples on random data to measure variability in archetype positions.
Set to NA fit once to all data instead of bootstraping.
When this option is positive seed bootstrap_seed and
sample_prop must be provided. (from ParetoTI)}

\item{result_name}{<character> default aa; name of used to store the
result in the SingleCellExperiment object.}

\item{envname}{<character> default r-decomp; specifies the name of the
python virtual environment to be used. If it does not exists it will
create one and use it.}

\item{return_model}{<bool> default FALSE; Whether to return also the model
and not only the SingleCellExperiment object.}

\item{seed}{<integer> default 42; to set the seed for reproducibility.}

\item{verbose}{<bool> default FALSE; Whether to be prompted with message
for each step of the analysis.}
}
\value{
either a SingleCellExperiment object with AA representation for both
genes and cells, or the SingleCellExperiment object and the model used to
perform archetypal analysis.
}
\description{
This function take has input a SingleCellExperiment object and performs
Archetypal Analysis (AA) leveraging (based on ParetoTI) on a
specified compressed representation.One can perform AA on the whole data
set, or on subsets of it via bootstrapping to account for possible outliers.
}
\examples{
#decomp_aa(sce = sce, n_components = 5, reduced_representation = 'pca')
}
