% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_ica.R
\name{decomp_ica}
\alias{decomp_ica}
\title{Decomposition Approach : Independent Component Analysis}
\usage{
decomp_ica(
  sce,
  n_components,
  assay = "logcounts",
  method = "fast",
  center = TRUE,
  scale = FALSE,
  result_name = "ica",
  return_model = FALSE,
  seed = 42,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sce}{<SingleCellExperiment object> SCE object}

\item{n_components}{<integer> number of Independent Component desired.}

\item{assay}{<character> assay used as input for ICA.}

\item{method}{<character> default **'fast'**; method used to perform
ICA. It can be one of (**"fast"**, **"imax"**, **"jade"**).
**'fast'** : Hyvarinen’s (1999) FastICA algorithm.
**'imax'** : Bell and Sejnowski’s (1995) Information-Maximization algorithm.
**'jade'** : Cardoso and Souloumiac’s (1993, 1996) Joint Approximate
Diagonalization of Eigenmatrices (JADE) algorithm.}

\item{center}{<bool> default TRUE; Whether to center the data or not;
centering is performed column wise, so on the cells (observations).}

\item{scale}{<bool> default FALSE; Whether to scale the data or not;
scaling is performed column wise, so on the cells (observations).}

\item{result_name}{<character> default 'ica';
Name of used to store the result in the SingleCellExperiment object.}

\item{return_model}{<bool> default FALSE; Whether to return also
the model and not only the SingleCellExperiment object.}

\item{seed}{<integer> default 42; to set the seed for reproducibility.}

\item{verbose}{<bool> default FALSE; Whether to be prompted with message
for each step of the analysis.}

\item{...}{<extra arguments for the specific implementation of ica used>}
}
\value{
either a SingleCellExperiment object with PCA representation for
only genes, or the SingleCellExperiment object and the model
used to perform PCA.
}
\description{
This function performs Independent Component Analysis (ICA) on a
SingleCellExperiment object and stores the results directly in it.
It is based on the ica package.
}
\examples{
library('packageX')
data(sce)
decomp_ica(sce, n_components = 6, assay = 'logcounts')
}
