% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_kaa.R
\name{decomp_kaa}
\alias{decomp_kaa}
\title{Decomposition Approach : Kernel Archetypal Analysis}
\usage{
decomp_kaa(
  sce,
  n_components,
  n_highly_variable_genes = 500,
  n_pcs = 50,
  kernel_dimension = "X_pca",
  min_iter = 10,
  max_iter = 10000,
  suggested_cells4aa = 75,
  n_waypoint_eigs = NULL,
  convergence_epsilon = 1e-05,
  minimum_weight = 0.05,
  use_highly_variable_genes = TRUE,
  summarize_layer = "raw",
  low_dim_embedding = "X_pca",
  plots = FALSE,
  bins = 5,
  celltype_label = NULL,
  nth_nearest_neighbor = 1,
  max_n_comps_pca = 80,
  kernel_matrix_plot_dim = seq(200),
  figures_dir = "kaa_plots",
  elbow_plot_name = "1_kaa_elbow_plot.pdf",
  kernel_matrix_plot_name = "2_portion_of_kernel_matrix.pdf",
  initialization_plot_name = "3_initialization_plot.pdf",
  convergence_plot_name = "4_convergence_plot.pdf",
  assignment_plot_name = "6_cell_assignment_plot.pdf",
  purity_plot_name = "7_archetypes'_purity_plot.pdf",
  compactness_plot_name = "8_archetypes'_compactness_plot.pdf",
  separation_plot_name = "9_archetypes'_separation_plot.pdf",
  result_name = "kaa",
  envname = "r-decomp",
  return_model = FALSE,
  seed = 42,
  verbose = TRUE
)
}
\arguments{
\item{sce}{<SingleCellExperiment object> SCE object}

\item{n_components}{<integer or NULL> number of archetypes desired.
Can be set to NULL to fallback to the
suggested number of archetypes (n_cells / suggested_cells4aa).}

\item{n_highly_variable_genes}{<integer> default 5000; number of
 Highly Variable Genes to be selected and then used to find the archetypes.
note: setting the HVGs number too low will cause the method to fail,
empirically you want to use 1500 genes or more as HVGs.
For smaller data set one may use roughly half of the total number of genes.}

\item{n_pcs}{<integer> default 50; number of Principal Components
used for downstream analysis.}

\item{kernel_dimension}{<charter> default 'X_pca'; representation of
the data used to build the kernel (it refers to the AnnData object).}

\item{min_iter}{<integer> default 10; minimum number of iterations to be
performed before convergence of the AA algorithm. If convergence is reached
before the minimum number of iterations it will not be considered.}

\item{max_iter}{<integer> default 10000; maximum number of iterations to be
performed before convergence of the AA algorithm. Once reached the method
will simply fail.}

\item{suggested_cells4aa}{<integer> default 75; Number of suggested cells
for a single archetype. 75 is the number suggested by the authors of the
core method (SEACells). If n_components is specified, suggested_cells4aa will
be ignored.}

\item{n_waypoint_eigs}{<integer> default NULL; if NULL it will fallback
to the suggested number of way points, 15. Way points are the number of
eigen-values to consider when initializing archetypes.
note: the user should never lower the number of way points further 3!}

\item{convergence_epsilon}{<double> default 1e-5; Convergence criteria.
A lower convergence_epsilon will result in more accurate solutions, but
also in a longer computation time.}

\item{minimum_weight}{<double> default 0.05; minimal threshold to consider
the contribution of an archetype to a cell. If the contribution is lower, it
will be set to 0!}

\item{use_highly_variable_genes}{<bool> default TRUE; Whether to use only
Highly Variable Genes to compute Archetypes.}

\item{summarize_layer}{<character> default 'raw'; Starting data matrix used
for computing archetypes. It refers to the AnnData object.}

\item{low_dim_embedding}{<character> default 'X_pca'; low dimensional
representation used to compute the compactness.}

\item{plots}{<bool> default FALSE; whether to plot quality metrics plot.}

\item{bins}{<integer> default 5; Number of bins in the plot of
kernel archetypes contributions to cells.}

\item{celltype_label}{<character> default NULL; if specified it will compute
archetype purity based on the defined field of the object. Generally is
cell_type or cluster… . Used for plotting quality metrics.}

\item{nth_nearest_neighbor}{<integer> default 1;
computes the diffusion distance from its nth nearest neighbor, used for
plotting quality metrics.}

\item{max_n_comps_pca}{<integer> default 80; maximum number of Principal
Components computed while performing PCA and visualized in the elbow plot.}

\item{kernel_matrix_plot_dim}{<integer vectors> default seq(200). Number of
cells to plot in the kernel matrix (it will be a squared matri). Used for
plotting quality metrics.}

\item{figures_dir}{<character> default 'kaa_plots'. Name of the directory
used to store all the quality metrics plots. If it does not exists it will
be created.}

\item{elbow_plot_name}{<character> default '1_kaa_elbow_plot.pdf';
Name of the elbow plot saved in figures_dir.}

\item{kernel_matrix_plot_name}{<character> default
'2_portion_of_kernel_matrix.pdf';
Name of the kernel matrix plot saved in figures_dir.}

\item{initialization_plot_name}{<character> default
'3_initialization_plot.pdf';
Name of the initialization plot saved in figures_dir.}

\item{convergence_plot_name}{<character> default '4_convergence_plot.pdf';
Name of the convergence plot saved in figures_dir.}

\item{assignment_plot_name, }{<character> default '6_cell_assignment_plot.pdf'
Name of the cell assignment plot saved in figures_dir.}

\item{purity_plot_name}{<character> default '7_archetypes\'_purity_plot.pdf'
Name of the archetypes' purity plot saved in figures_dir.}

\item{compactness_plot_name}{default '8_archetypes\'_compactness_plot.pdf';
Name of the archetypes' compactness plot saved in figures_dir.}

\item{separation_plot_name}{default '9_archetypes\'_separation_plot.pdf';
Name of the archetypes' separation plot saved in figures_dir.}

\item{result_name}{<character> default 'kaa';
Name used to store the result in the SingleCellExperiment object.}

\item{envname}{<character> default 'r-decomp';
Specify the name of the python virtual
environment to be used. If it does not exists it will create one and use it.}

\item{return_model}{<bool> default FALSE; Whether to return also
the model and not only
the SingleCellExperiment object.}

\item{seed}{<integer> default 42; to set the seed for reproducibility.}

\item{verbose}{<bool> default FALSE; Whether to be prompted with message
for each step of the analysis.}
}
\value{
either a SingleCellExperiment object with PCA representation for
only genes, or the SingleCellExperiment object and the model
used to perform PCA.
}
\description{
This function performs Kernel Archetypal Analysis leveraging SEACells
implementation and store directly in the SingleCellExperiment object the
results.
}
\examples{
library('packageX')
data(sce)
decomp_kaa(sce, n_components = 5, n_highly_variable_genes = 250, n_pcs = 10, n_waypoint_eigs = 3, verbose = TRUE, plots = TRUE)
}
