% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_kb.R
\name{decomp_kb}
\alias{decomp_kb}
\title{Decomposition Approach : Knowledge Based Decomposition}
\usage{
decomp_kb(
  sce,
  metadata_key,
  prior_knowledge,
  n_components = NULL,
  assay = "logcounts",
  lam = 0.1,
  n_hvgs = 5000,
  n_markers = 50,
  num_epochs = 10000,
  min_gs_num = 3,
  use_metadata = TRUE,
  use_weights = TRUE,
  delta = 0.001,
  gene_set_confidence_assignment = 0.2,
  additional_results = TRUE,
  ignore_warnings = TRUE,
  ggg_metadata = NULL,
  result_name = "kb",
  envname = "r-decomp",
  return_model = FALSE,
  seed = 42,
  verbose = FALSE
)
}
\arguments{
\item{sce}{<SingleCellExperiment object> SCE object}

\item{metadata_key}{<'character or NULL'> default NULL; Specified prior
knowledge to be used, specifically groups of cells (like clusters, types,
states, …). If is set to NULL, prior_knowledge should not have any metadata
specific gene set!}

\item{prior_knowledge}{<list of lists of named gene sets> default NULL;
List of lists, where each sub list is a named gene set to be tested.
Gene sets can be metadata specific or global. If global the gene set will
be tested on all the cells, regardless of the grouping due to the metadata.
The gene sets should be grouped in a named list; the name of the list must
match a name of the metadata_key.
look at the example to get an idea of the structure…}

\item{n_components}{<integer or NULL> default NULL. Generally should match
the number of gene sets (authors suggestion).}

\item{assay}{<character> default 'logcounts'; assay used for the analysis.}

\item{lam}{<double> default 0.1; trust in prior knowledge.}

\item{n_hvgs}{<integer or NULL> default 5000; Number of Highly Variable Genes
to be selected and used for downstream analysis. If set to 0 or NULL, it will
use all the available genes.}

\item{n_markers}{<integer> default 50; number of genes for each gene set.
It accounts also for imputed genes during the analysis based on the data.}

\item{num_epochs}{<integer> default 10000; number of epocs used to fit
the model. 10.000 is the recomended default.}

\item{min_gs_num}{<integer> default 3; minimum number of gene in a gene set.}

\item{use_metadata}{<bool> default TRUE; Whether to use the metadata_key
or not. If set to FALSE, the user should not specify metadata specific gene
sets.}

\item{use_weights}{<bool> default TRUE; edge weights are estimated based
on graph structure and used throughout training (from Spectra).}

\item{delta}{<double> default 0.001; Scaling parameter to make the values
comparable. It varies depending on data and gene sets.
It can range from 0.5 up to 0.001.}

\item{gene_set_confidence_assignment}{<double> default 0.02;
minimum fraction of overlapping genes to assign the gene set to a
given metadata group of cells.}

\item{additional_results}{<bool> default TRUE; Whether to store in the
metadata field of the SingleCellExperiment object also Markers of each Set,
Interpretation of Factors, Influence of the Prior on the Factors and the
Gene-Gene graph.}

\item{ignore_warnings}{<bool> default TRUE; Whether to silence warnings.}

\item{ggg_metadata}{<character or NULL>; Specifying a metadata
group (like a given cell type) to obtain a Metadata-Specific Gene-Gene Graph,
thus the name ggg_metadata.}

\item{result_name}{<character> default 'kb';
Name of used to store the result in the SingleCellExperiment object.}

\item{envname}{<character> default 'r-decomp';
Specify the name of the python virtual
environment to be used. If it does not exists it will create one and use it.}

\item{return_model}{<bool> default FALSE; Whether to return also
the model and not only
the SingleCellExperiment object.}

\item{seed}{<integer> default 42; to set the seed for reproducibility.}

\item{verbose}{<bool> default FALSE; Whether to be prompted with message
for each step of the analysis.}
}
\value{
either a SingleCellExperiment object with PCA representation for
only genes, or the SingleCellExperiment object and the model
used to perform PCA.
}
\description{
This function performs Knowledge Based Decomposition leveraging Spectra
implementation. It stores the results directly inside the
SingleCellExperiment object.
}
\examples{
# prior_knowledge = list('0' = list('gene_set_1' = list("KLHL17",
#                                                      "CCNL2",
#                                                      "ATAD3B",
#                                                      "NOL9"),
#                                  'gene_set_2' = list("CENPS-CORT",
#                                                      "FBXO2",
#                                                      "KLHL21")),
#                        '1' = list('gene_set_1' = list("NPPA",
#                                                       "FHAD1",
#                                                       "FBXO42",
#                                                       "ATP13A2"),
#                                   'gene_set_3' = list("PADI2",
#                                                       "AHDC1",
#                                                       "PPP1R8",
#                                                       "LAPTM5")),
#                  # set of genes which will be evaluated for all the
#                  # cells in data…
#                  'global' = list('gene_set_4' = list("AZIN2",
#                                                      "ZSCAN20",
#                                                      "ZC3H12A")))
# decomp_kb(sce,
#           metadata_key = 'is.doublet',
#           prior_knowledge = prior_knowledge,
#           n_hvgs = NULL,
#           num_epochs = 100)
}
