% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_mraa.R
\name{decomp_mraa}
\alias{decomp_mraa}
\title{Decomposition Approach : MultiResolution Archetypal Analysis}
\usage{
decomp_mraa(
  sce,
  n_components = NULL,
  norm.assay = NULL,
  reduced.assay = NULL,
  store_all_results = FALSE,
  result_name = "mraa",
  return_model = FALSE,
  seed = 42,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sce}{<SingleCellExperiment object> SCE object}

\item{n_components}{<integer or NULL> default NULL; number of archetypes
desired. If specified it will force a single resolution, the number of
specified archetypes. Otherwise, it will work at multiple resolution and will
prune useless or repeated archetypes found at multiple resolution, keeping
only informative ones.}

\item{norm.assay}{<character or NULL> default NULL; name of the assay which
stores the normalized counts. If NULL it will be obtained via shifted
logarithm normalization.}

\item{reduced.assay}{<character or NULL> default NULL; name of the assay
which stores the dimensionality reduction representation of H
(cells x latents). If NULL it will be obtained via randomized PCA.}

\item{store_all_results}{<bool> default FALSE; Whether to store all the
intermediate results obtained via ACTIONet in the SingleCellExperiment
object.}

\item{result_name}{<character> default 'mraa';
Name used to store the result in the SingleCellExperiment object.}

\item{return_model}{<bool> default FALSE; Whether to return also
the model and not only
the SingleCellExperiment object.}

\item{seed}{<integer> default 42; to set the seed for reproducibility.}

\item{verbose}{<bool> default FALSE; Whether to be prompted with message
for each step of the analysis.}

\item{...}{<extra arguments for the runACTIONet function>}
}
\value{
either a SingleCellExperiment object with H and W representation
obtained via MultiResolution Archetypal Analysis, or the SingleCellExperiment
and the ACTIONet object used to perform the various steps
}
\description{
This function performs MultiResolution Archetypal Analysis as implemented by
the ACTIONet framework. It stores directly in the SingleCellExperiment object
the results. It can work in parallel using multiple cores.
}
\examples{
library(packageX)
data(sce)
#decomp_mraa(sce, n_components = 14, verbose = TRUE)
}
