% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_sparse_pca.R
\name{decomp_sparse_pca}
\alias{decomp_sparse_pca}
\title{Decomposition Approach : Sparse-Matrix Principal Component Analysis}
\usage{
decomp_sparse_pca(
  sce,
  n_components = 50,
  assay = "logcounts",
  center = TRUE,
  scale = TRUE,
  result_name = "pca",
  return_model = FALSE,
  seed = 42,
  verbose = FALSE
)
}
\arguments{
\item{sce}{<SingleCellExperiment object> SCE object}

\item{n_components}{<integer> number of Principal Components desired.
Must be n_components == min(n_cells, n_genes)}

\item{assay}{<character> default logcounts;
specifying the assay to use, preferred raw counts!}

\item{center}{<bool> default TRUE;
Whether to center in a column wise fashion
the assay prior to performing PCA.}

\item{scale}{<bool> default TRUE;
Whether to scale in a column wise fashion
the assay prior to performing PCA.}

\item{result_name}{<character> default 'pca';
Name of used to store the result in the SingleCellExperiment object.}

\item{return_model}{<bool> default FALSE; Whether to return also
the model and not only
the SingleCellExperiment object.}

\item{seed}{<integer> default 42; to set the seed for reproducibility.}

\item{verbose}{<bool> default FALSE; Whether to be prompted with message
for each step of the analysis.}
}
\value{
either a SingleCellExperiment object with PCA representation for
only genes, or the SingleCellExperiment object and the model
used to perform PCA.
}
\description{
This function performs Principal Component Analysis (PCA) keeping a
sparse representation for the matrix and leveraging irlba::prcomp_irlba
implementation.
}
\examples{
#decomp_sparse_pca(sce, n_components = 50)
}
