% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hard_assignment.R
\name{hard_assignment}
\alias{hard_assignment}
\title{Hard Assignment of Latents to an Observation (row)}
\usage{
hard_assignment(matrix, margin = 1)
}
\arguments{
\item{matrix}{<dgCMatrix or matrix generally> The matrix H (observations x
latents) which will be used for the hard assignment.}

\item{margin}{<either 1 or 2> dimension used to pick the maximum, 1 to
work on the latents, so the rows!}
}
\value{
Returns a vector with the assigned latent for each cell.
}
\description{
This function performs hard assignment of latents (columns) to
observations (rows) by using the highest weight as criteria of assignment.
}
\examples{
library(packageX)
data(sce)
hard_assignment(SingleCellExperiment::reducedDim(sce, 'pca'))
}
