% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignore_warnings.R
\name{ignore_warnings}
\alias{ignore_warnings}
\title{Ignore Warnings}
\usage{
ignore_warnings(ignore_warnings, envname = "r-decomp", verbose = FALSE)
}
\arguments{
\item{ignore_warnings}{<bool> Whether to ignore warning related to the
python environment.}

\item{envname}{<character> default 'r-decomp';
Specify the name of the python virtual
environment to be used. If it does not exists it will create one and use it.}

\item{verbose}{<bool> default FALSE; Whether to be prompted with message
for each step of the analysis.}
}
\value{
nothing, it is a simple side-effect function.
}
\description{
This function sets the python environment to ignore warnings if provided
with ignore_warning = TRUE.
}
\examples{
#ignore_warnings(TRUE, verbose = TRUE)
}
