% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_MPS_avaliable.R
\name{is_MPS_avaliable}
\alias{is_MPS_avaliable}
\title{Checks if MPS Acceleration is Available}
\usage{
is_MPS_avaliable(envname, verbose)
}
\arguments{
\item{envname}{<character> default 'r-decomp';
Specify the name of the python virtual
environment to be used. If it does not exists it will create one and use it.}

\item{verbose}{<bool> default FALSE; Whether to be prompted with message
for each step of the analysis.}
}
\value{
nothing, it is a simple side-effect function. If verbose is set to
TRUE, it will print a message to the end user stating whether it could
use MPS acceleration or not.
}
\description{
This function checks if the MPS acceleration is available. It is
participatory useful when working with Neural Networks like a Variational
AutoEncoder.
}
\examples{
is_MPS_avaliable(envname = 'r-decomp', verbose = TRUE)
}
