% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_package_installed.R
\name{is_package_installed}
\alias{is_package_installed}
\title{Checks if an R Package is Available}
\usage{
is_package_installed(package_x, lib.loc = NULL)
}
\arguments{
\item{package_x}{<character> Name of the needed package}

\item{lib.loc}{<NULL or character> default NULL; Path to the R library.
If NULL, it will use the default R library. Instead if specified it will
use the one specified. This could be particular useful when working on
a shared R library and a user specific R library in a High Performance
Computing (HPC).}
}
\value{
nothing, it is a simple side-effect function. It will break the
function if the package is not available. But it will fail gracefully.
}
\description{
Handy function to check if an R package can be used or not.
If the specified R package is nowhere to be found it will fail
gracefully.
}
