% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_assignment_matrix.R
\name{list_to_assignment_matrix}
\alias{list_to_assignment_matrix}
\title{It transforms a List into an Assignment Matrix}
\usage{
list_to_assignment_matrix(input.list)
}
\arguments{
\item{input.list}{<list or vector> the input list which will be transformed
into an assignment matrix}
}
\value{
a sparse (dgCMatrix) assignment matrix
}
\description{
This function takes a list and returns a sparse assignment matrix.
}
\examples{
list_to_assignment_matrix(input.list = c('a', 'a', 'b', 'c'))
}
