% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_warning.R
\name{package_warning}
\alias{package_warning}
\title{Handy function for Warnings Messages}
\usage{
package_warning(package_x)
}
\arguments{
\item{package_x}{<character> the name of the package need to be installed}
}
\value{
a nice message providing instruction on how to install it
}
\description{
This function takes the name of an R package we want to install and returns
a nice message warning on how to install it. It is a general message, but
it makes for a graceful break if the package has not been installed.
}
\examples{
#package_warning('NotToBeFoundPackage')
}
