% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_graph.R
\name{partition_graph}
\alias{partition_graph}
\title{Partition Approach : Graph Clustering}
\usage{
partition_graph(
  sce,
  resolution = 1,
  method = "leiden",
  n_pcs = 50,
  n_neighbors = 10,
  svd_solver = "auto",
  ignore_warnings = TRUE,
  result_name = "leiden",
  envname = "r-partition",
  return_model = FALSE,
  seed = 42,
  verbose = FALSE
)
}
\arguments{
\item{sce}{<SingleCellExperiment object> SCE object}

\item{resolution}{<double> default 1; resolution used for clustering.
The higher it is, the finer the clusters; The lower it is, the more coarser
the clusters.}

\item{method}{<character> default *leiden*. Method used for clustering, can
be one of:
*louvain* : modularity based clustering;
*leiden*  : modularity based clustering with significant improvement over
louvain.}

\item{n_pcs}{<integer> default 50; number of PCs to be used after Principal
Component Analysis as reduction method.}

\item{n_neighbors}{<integer> default 10; number of Nearest Neighbors to be
used for the K-Nearest Neighbors (KNN) Graph.}

\item{svd_solver}{<character> default *auto*; Type of solver to perform
Principal Component Analysis. Can be one of :
*arpack*    : default solver form the ARPACK wrapper in SciPy;
*randomized*: for a randomized PCA (Halko - 2009);
*auto*      : the solver is picked depending on the type of data used;
*lobpcg*    : experimental solver to process quickly a sparse matrix.
note : *lobpcg* currently (Q1-2024) only works with the ‘arpack’
or ‘lobpcg’ solvers.}

\item{ignore_warnings}{<bool> Whether to ignore warning or not.}

\item{result_name}{<character> default 'leiden';
Name used to store the result in the SingleCellExperiment object.}

\item{envname}{<character> default 'r-partition';
Specify the name of the python virtual
environment to be used. If it does not exists it will create one and use it.}

\item{return_model}{<bool> default FALSE; Whether to return also
the model and not only
the SingleCellExperiment object.}

\item{seed}{<integer> default 42; to set the seed for reproducibility.}

\item{verbose}{<bool> default FALSE; Whether to be prompted with message
for each step of the analysis.}
}
\value{
either a SingleCellExperiment object with both H and W
representations or the SingleCellExperiment object and the model
used to perform X
}
\description{
This function performs Graph Clustering via either Louvain or Leiden on a
SingleCellExperiment (SCE) object. It stores the memebership vector,
the result of the function, directly in the colData slot of the SCE Object.
}
\examples{
#partition_graph(sce)
}
