% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aas_feature.R
\name{plot_aas_feature}
\alias{plot_aas_feature}
\title{Handy Function for select_aas for Quality Metrics.}
\usage{
plot_aas_feature(
  aas.model,
  type,
  title,
  point_size = 0.5,
  line_size = 0.5,
  text_axis_size = 10,
  plot_title_size = 15,
  hjust = 0.5
)
}
\arguments{
\item{aas.model}{<model of select_aas> model obtained from select_aas()}

\item{type}{<character> type of quality metric to be plotted.
example : *total_var*.}

\item{title}{<character> title for the plot.
example : *Archetypes Variability*}

\item{point_size}{<double> default 0.5; dimensions of the plotted point.}

\item{line_size}{<double> default 0.5; width of the plotted lines.}

\item{text_axis_size}{<integer or double> default 10; dimension of the text
on the axis.}

\item{plot_title_size}{<integer or double> default 15; dimension of
the title.}

\item{hjust}{<double> default 0.5; Position of the title (by default in the
central).}
}
\value{
a Quality Metric plot defined by the type variable.
}
\description{
This function provides quality metrics plots for select_aas model with .
}
