% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_model.R
\name{return_model}
\alias{return_model}
\title{Handy Function to Return Multiple Objects}
\usage{
return_model(sce, model, return_model, ...)
}
\arguments{
\item{sce}{<SingleCellExperiment object> SCE object}

\item{model}{<object of various types> model used in a
partition or decomposition method.}

\item{return_model}{<bool> default FALSE; Whether to return also
the model and not only the SingleCellExperiment object.}

\item{...}{<extra arguments to be returned, must be specified as
new_argumnet_name = new_argument>}
}
\value{
either a SCE object or the SCE object and the model used to perform
the decompostion or partition method
}
\description{
This function lets you return 2 different object, the SingleCellExperiment
(SCE) object and the model used during the analysis. It may return only the
SCE object or both the SCE and the model given the return_model boolean
variable.
}
\examples{
#return_model(sce = c(1, 2 ,3), model = c('a', 'b', 'c'), return_model = TRUE, bool = c(TRUE, FALSE))
}
