% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_pdf.R
\name{save_pdf}
\alias{save_pdf}
\title{Handy Function to Save a ComplexHeatmap in .pdf format}
\usage{
save_pdf(plot, file_name, width = 5, height = 3.25)
}
\arguments{
\item{plot}{ComplexHeatmap}

\item{file_name}{<character.pdf> name of the file used to store the plot.}

\item{width}{<double> default 5;
Width of the plot}

\item{height}{<double> default 3.5;
height of the plot}
}
\value{
a .pdf file of a ComplexHeatmap plot
}
\description{
This function saves a ComplexHeatmap in .pdf format
}
\examples{
save_pdf(plot = ComplexHeatmap::Heatmap(seq(1, 10, length.out = 10), col = grDevices::blues9), file_name = '~/Documents/plot.pdf')
}
