% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce}
\alias{sce}
\title{Simulated Brain Organoids single-cell RNA-seq data}
\format{
## `sce`
A data frame with 500 rows -genes;features-
and 1000 columns -cells; observations-:
\describe{
  \item{cluster}{integer from 1 to 13; Identity programs of the cells}
  \item{program.usage}{double from 0 up to 1; Usage of the state program}
  \item{is.doublet}{1 or 0; Wether the cell is a doublet or not}
  ...
}
}
\source{
<https://codeocean.com/capsule/6314882/tree/v1>
}
\usage{
sce
}
\description{
A subset of data from https://github.com/dylkot/cNMF
They present 13 different Identity communities and 1 shared state phenotype.
note: the gene names where picked at random from another data set
}
\keyword{datasets}
