% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce2adata_sparse.R
\name{sce2adata_sparse}
\alias{sce2adata_sparse}
\title{Handy Function to Convert a SingleCellExperiment Object into AnnData Object}
\usage{
sce2adata_sparse(sce, envname = "r-decomp", main_layer = "counts", ...)
}
\arguments{
\item{sce}{<SingleCellExperiment object> SCE object}

\item{envname}{<character> default 'r-decomp';
Specify the name of the python virtual
environment to be used. If it does not exists it will create one and use it.}

\item{main_layer}{<character> default counts. Main layer to be converted.}

\item{...}{<extra arguments for the function convertFormat>}
}
\value{
an AnnData object with most of the fields of the original SCE
}
\description{
This function converts a SingleCellExperiment (SCE) object into AnnData
object and it keeps most of the fields intact! It is also very useful in
converting the matrix types in a proper way.
}
\examples{
library(packageX)
data(sce)
sce2adata_sparse(sce)
}
