% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_sparse.R
\name{split_sparse}
\alias{split_sparse}
\title{Splits a Sparse Matrix (without converting it into a dense one)}
\usage{
split_sparse(sparse_matrix, split_vector, by_col = TRUE, sorted = TRUE)
}
\arguments{
\item{sparse_matrix}{<matrix, better if in sparse format> The matrix to be
split.}

\item{split_vector}{<vector> vector used to split the sparse_matrix into
smaller ones. The split is by default by column (by_col = TRUE); Hence the
split_vector must be the same length as the number of columns.}

\item{by_col}{<bool> default 'X'; Whether to split the matrix by columns or
by rows.}

\item{sorted}{<bool> default TRUE; Whether to sort the output or not.
This parameter is useful when dealing with a numeric split_vector, since the
if sorted = FALSE, the order of the split vector will define the order of
the output matrices.}
}
\value{
A list with many smaller matrices, one for each unique element in
the split_vector. The order of the matrices depends on the sorted variable.
}
\description{
This function splits a sparse matrix (generally dgCMatrix) without coercing
it into a dense representation.
}
\examples{
library(packageX)
data(sce)
split_sparse(sparse_matrix = SingleCellExperiment::counts(sce), split_vector = sce$cluster)
}
