% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_H.R
\name{store_H}
\alias{store_H}
\title{Handy Function to Store the H matrix in a SingleCellExperiment Object}
\usage{
store_H(sce, h.matrix, result_name, latent_name)
}
\arguments{
\item{sce}{<SingleCellExperiment object> SCE object}

\item{h.matrix}{<matrix generally dense> H matrix (cells x latents) to be
stored in the SCE object.}

\item{result_name}{<character> Name used to store the result in the
SCE object.}

\item{latent_name}{<character> Name used for the latents. It will be used
as column name followed by the number of the latent. (example AA; AA1 AA2 …)}
}
\value{
The SCE object with the new H representation stored in the
reducedDim slot.
}
\description{
This function stores a dense H matrix into a SingleCellExperiment (SCE)
object in the reducedDim slot in a sparse format (dgCMatrix). It also
adds the names to the rows and changes the name of the stored H matrix
if it is already present that name inside reducedDim slot.
}
\examples{
library(packageX)
data(sce)
store_H(sce, h.matrix = Matrix::t(matrix(stats::rnorm(2 * ncol(sce)), ncol = ncol(sce))), result_name = 'aa', latent_name = 'AA')
}
