% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_W.R
\name{store_W}
\alias{store_W}
\title{Handy Function to Store the W matrix in a SingleCellExperiment Object}
\usage{
store_W(sce, w.matrix, latent_name, result_name)
}
\arguments{
\item{sce}{<SingleCellExperiment object> SCE object}

\item{w.matrix}{<matrix generally dense> w matrix (genes x latents) to be
stored in the SCE object.}

\item{latent_name}{<character> Name used for the latents. It will be used
as column name followed by the number of the latent. (example AA; AA1 AA2 …)}

\item{result_name}{<character> Name used to store the result in the
SCE object.}
}
\value{
The SCE object with the new W representation stored in the
metadata slot.
}
\description{
This function stores a dense W matrix into a SingleCellExperiment (SCE)
object in the metadata slot in a sparse format (dgCMatrix). It also
adds the names to the rows and changes the name of the stored W matrix
if it is already present that name inside the reducedDimNames slot.
Hence you want to use the store_W before the store_H function!
}
\examples{
#store_W(sce, w.matrix =  Matrix::t(matrix(stats::rnorm(2 * nrow(sce)), nrow = nrow(sce))), result_name = 'aa', latent_name = 'AA')
}
